+++ examples/string.cpp
@@ -14,7 +14,6 @@
  */
 
 #include <aws/cryptosdk/cpp/kms_keyring.h>
-#include <aws/cryptosdk/default_cmm.h>
 #include <aws/cryptosdk/enc_ctx.h>
 #include <aws/cryptosdk/session.h>
 
@@ -33,34 +32,23 @@ int encrypt_string(
         return 2;
     }
 
-    struct aws_cryptosdk_cmm *cmm = aws_cryptosdk_default_cmm_new(alloc, kms_keyring);
-    if (!cmm) {
-        /* Any new function can return NULL on a memory allocation failure.
-         * We give up on those.
-         */
-        aws_cryptosdk_keyring_release(kms_keyring);
-        return 3;
-    }
+    struct aws_cryptosdk_session *session =
+        aws_cryptosdk_session_new_from_keyring(alloc, AWS_CRYPTOSDK_ENCRYPT, kms_keyring);
 
-    /* The CMM has a reference to the keyring now. We release our reference so that
-     * the keyring will be destroyed when the CMM is.
+    /* The session has a reference to the keyring now. We release our reference so that
+     * the keyring will be destroyed when the session is. If session failed to allocate,
+     * this causes an immediate deallocation of the keyring and prevents a memory leak
+     * when this function exits.
      */
     aws_cryptosdk_keyring_release(kms_keyring);
 
-    struct aws_cryptosdk_session *session = aws_cryptosdk_session_new_from_cmm(alloc, AWS_CRYPTOSDK_ENCRYPT, cmm);
     if (!session) {
-        aws_cryptosdk_cmm_release(cmm);
-        return 4;
+        return 3;
     }
-
-    /* The session has a reference to the CMM now. We release our reference so that
-     * when the CMM (and the keyring) will be destroyed when the session is.
-     */
-    aws_cryptosdk_cmm_release(cmm);
 
     if (AWS_OP_SUCCESS != aws_cryptosdk_session_set_message_size(session, plaintext_len)) {
         aws_cryptosdk_session_destroy(session);
-        return 5;
+        return 4;
     }
 
     /* The encryption context is an AWS hash table where both the key and value
@@ -77,7 +65,7 @@ int encrypt_string(
     /* We copy the contents of our own encryption context into the session's. */
     if (AWS_OP_SUCCESS != aws_cryptosdk_enc_ctx_clone(alloc, session_enc_ctx, my_enc_ctx)) {
         aws_cryptosdk_session_destroy(session);
-        return 6;
+        return 5;
     }
 
     /* We encrypt the data. */
@@ -86,17 +74,17 @@ int encrypt_string(
         aws_cryptosdk_session_process(
             session, ciphertext, ciphertext_buf_sz, ciphertext_len, plaintext, plaintext_len, &plaintext_consumed)) {
         aws_cryptosdk_session_destroy(session);
-        return 7;
+        return 6;
     }
 
     if (!aws_cryptosdk_session_is_done(session)) {
         aws_cryptosdk_session_destroy(session);
-        return 8;
+        return 7;
     }
     if (plaintext_consumed != plaintext_len) abort();
 
     /* This call deallocates all of the memory allocated in this function, including
-     * the keyring and CMM, since we already released their pointers.
+     * the keyring, since we already released its pointer.
      */
     aws_cryptosdk_session_destroy(session);
     return 0;
@@ -114,36 +102,28 @@ int decrypt_string_and_verify_encryption_context(
     struct aws_cryptosdk_keyring *kms_keyring = Aws::Cryptosdk::KmsKeyring::Builder().Build(key_arn);
     if (!kms_keyring) {
         fprintf(stderr, "Failed to build KMS Keyring. Did you specify a valid KMS CMK ARN?\n");
-        return 9;
+        return 8;
     }
 
-    struct aws_cryptosdk_cmm *cmm = aws_cryptosdk_default_cmm_new(alloc, kms_keyring);
-    if (!cmm) {
-        aws_cryptosdk_keyring_release(kms_keyring);
-        return 10;
-    }
-
+    struct aws_cryptosdk_session *session =
+        aws_cryptosdk_session_new_from_keyring(alloc, AWS_CRYPTOSDK_DECRYPT, kms_keyring);
     aws_cryptosdk_keyring_release(kms_keyring);
 
-    struct aws_cryptosdk_session *session = aws_cryptosdk_session_new_from_cmm(alloc, AWS_CRYPTOSDK_DECRYPT, cmm);
     if (!session) {
-        aws_cryptosdk_cmm_release(cmm);
-        return 11;
+        return 9;
     }
-
-    aws_cryptosdk_cmm_release(cmm);
 
     size_t ciphertext_consumed;
     if (AWS_OP_SUCCESS !=
         aws_cryptosdk_session_process(
             session, plaintext, plaintext_buf_sz, plaintext_len, ciphertext, ciphertext_len, &ciphertext_consumed)) {
         aws_cryptosdk_session_destroy(session);
-        return 12;
+        return 10;
     }
 
     if (!aws_cryptosdk_session_is_done(session)) {
         aws_cryptosdk_session_destroy(session);
-        return 13;
+        return 11;
     }
     if (ciphertext_consumed != ciphertext_len) abort();
 
@@ -155,16 +135,15 @@ int decrypt_string_and_verify_encryption_context(
     const struct aws_hash_table *session_enc_ctx = aws_cryptosdk_session_get_enc_ctx_ptr(session);
     if (!session_enc_ctx) abort();
 
-    /* Because the CMM can add new entries to the encryption context, we do not
-     * require that the encryption context matches, but only that the entries we
-     * put in are there.
+    /* Because the AWS Encryption SDK can add new entries to the encryption context, we do not
+     * require that the encryption context matches, but only that the entries we put in are there.
      */
     for (struct aws_hash_iter iter = aws_hash_iter_begin(my_enc_ctx); !aws_hash_iter_done(&iter);
          aws_hash_iter_next(&iter)) {
         struct aws_hash_element *session_enc_ctx_kv_pair;
         if (AWS_OP_SUCCESS != aws_hash_table_find(session_enc_ctx, iter.element.key, &session_enc_ctx_kv_pair)) {
             aws_cryptosdk_session_destroy(session);
-            return 14;
+            return 12;
         }
 
         if (!session_enc_ctx_kv_pair ||
@@ -181,7 +160,7 @@ int decrypt_string_and_verify_encryption_context(
 
 /* Allocates a hash table for holding the encryption context and puts a few sample values in it. */
 int set_up_enc_ctx(struct aws_allocator *alloc, struct aws_hash_table *enc_ctx) {
-    if (AWS_OP_SUCCESS != aws_cryptosdk_enc_ctx_init(alloc, enc_ctx)) return 12;
+    if (AWS_OP_SUCCESS != aws_cryptosdk_enc_ctx_init(alloc, enc_ctx)) return 13;
 
     /* Declares AWS strings of type (static const struct aws_string *)
      *
@@ -197,12 +176,12 @@ int set_up_enc_ctx(struct aws_allocator *alloc, struct aws_hash_table *enc_ctx) 
     int was_created;
     if (AWS_OP_SUCCESS != aws_hash_table_put(enc_ctx, enc_ctx_key1, (void *)enc_ctx_value1, &was_created)) {
         aws_cryptosdk_enc_ctx_clean_up(enc_ctx);
-        return 15;
+        return 14;
     }
     if (was_created != 1) abort();
     if (AWS_OP_SUCCESS != aws_hash_table_put(enc_ctx, enc_ctx_key2, (void *)enc_ctx_value2, &was_created)) {
         aws_cryptosdk_enc_ctx_clean_up(enc_ctx);
-        return 16;
+        return 15;
     }
     if (was_created != 1) abort();
     return 0;