+++ include/aws/cryptosdk/session.h
@@ -65,6 +65,22 @@ enum aws_cryptosdk_mode { AWS_CRYPTOSDK_ENCRYPT = 0x9000, AWS_CRYPTOSDK_DECRYPT 
  *                  data allocated for the session
  * @param mode The mode (AWS_CRYPTOSDK_ENCRYPT or AWS_CRYPTOSDK_DECRYPT) to start
  *             in. This can be changed later with @ref aws_cryptosdk_session_reset
+ * @param keyring The keyring which will encrypt or decrypt data keys for this session.
+ *                This function uses a default CMM to link the session and keyring.
+ */
+AWS_CRYPTOSDK_API
+struct aws_cryptosdk_session *aws_cryptosdk_session_new_from_keyring(
+    struct aws_allocator *allocator, enum aws_cryptosdk_mode mode, struct aws_cryptosdk_keyring *keyring);
+
+/**
+ * Creates a new encryption or decryption session.
+ *
+ * @return The new session, or NULL on failure (in which case, an AWS error code is set)
+ *
+ * @param allocator The allocator to use for the session object and any temporary
+ *                  data allocated for the session
+ * @param mode The mode (AWS_CRYPTOSDK_ENCRYPT or AWS_CRYPTOSDK_DECRYPT) to start
+ *             in. This can be changed later with @ref aws_cryptosdk_session_reset
  * @param cmm The crypto material manager which will provide key material for this
  *            session.
  */
@@ -190,7 +206,7 @@ bool aws_cryptosdk_session_is_done(const struct aws_cryptosdk_session *session);
  * returns true, but may succeed earlier in the message as well.
  */
 AWS_CRYPTOSDK_API
-int aws_cryptosdk_session_get_algorithm(const struct aws_cryptosdk_session *session, enum aws_cryptosdk_alg_id *alg_id);
+int aws_cryptosdk_session_get_alg_id(const struct aws_cryptosdk_session *session, enum aws_cryptosdk_alg_id *alg_id);
 
 /**
  * Estimates the amount of buffer space needed to make forward progress.